Security improvements, crypto unit tests and fuzz tests
    * LeastAuthority issue #3 - Implemented fuzz tests
        - Added validation that user-supplied Mina addresses are
          correct before signing transactions
        - Improved safety of generate_address and simplified error handling
        - Removed assertions from random_oracle_input and added error handling
        - Simplified and improved gen_address() error handling
        - Bugfix to length calculations and representations in APDU protocol
        - Security improvements to roinput_derive_message error handling and
          clearing secret key from stack
        - Improved error handling in read_public_key_compressed()
        - Replaced assertions with error handling for message_hash()

    * Security fix to Ledger boilerplate (do not trust user-supplied
      command data length!)

    * LeastAuthority issue #3 - Implemented crypto unit tests
        - Check that addition is commutative and associative; scaling
          commutes with the addition, multiplication and negation of
          scalars.  Check all computed points are on the curve and
          that the points match the reference implementation
        - Implemented group negation and affine_{eq,add,negate,is_on_curve}
        - Updated group scalar mul to not skip the first two bits
        - Switched to more consistent use of field_copy, scalar_copy
          and os_memcpy
        - Imported curve_checks() unit tests and constants from
          reference implementation

    * Fixed hidden stack exhaustion in sign(), added TRY/FINALLY error
      handling, clear sensitive data
        - Reduced stack consumption of group_add
        - Reduced stack consumption of affine_from_group
        - Reduced stack consumption of message_derive

    * Swapped to Ledger-recommended explicit_bzero() for clearing
      secrets and sensitive data

    * Updated documentation

+++ src/random_oracle_input.h
@@ -27,5 +27,5 @@ void roinput_add_bit(ROInput *input, bool b);
 void roinput_add_bytes(ROInput *input, const uint8_t *bytes, size_t len);
 void roinput_add_uint32(ROInput *input, const uint32_t x);
 void roinput_add_uint64(ROInput *input, const uint64_t x);
-size_t roinput_derive_message(uint8_t *out, size_t len, const Keypair *kp, const ROInput *msg);
-size_t roinput_hash_message(Field *out, size_t len, const Affine *pub, const Field rx, const ROInput *msg);
+int roinput_derive_message(uint8_t *out, size_t len, const Keypair *kp, const ROInput *msg);
+int roinput_hash_message(Field *out, size_t len, const Affine *pub, const Field rx, const ROInput *msg);
